requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-utils
greenlet==3.0.3
h11==0.14.0
idna==3.6
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.0.2
pytest-mock==3.12.0
pydantic==2.6.3